Actualizacion repo- cambio Reporte Alert

+++ hydrometrics-main-back/src/main/java/com/service/hydrometrics/models/DTO/user/CreateUserDTO.java
@@ -22,5 +22,5 @@ public class CreateUserDTO implements Serializable {
     private final String email;
     @NotEmpty
     private final String role;
-    private final byte enabled;
+    private final int enabled;
 }
